Rename InternalLink to StyledLink and deletd ExternalLink
