+++ src/components/Footer.js
@@ -2,23 +2,23 @@ import React from 'react';
 
 import logo from '../assets/logo-symbol.svg';
 import FooterContainer from './FooterContainer';
-import InternalLink from './InternalLink';
-import ExternalLink from './ExternalLink';
+import StyledLink from './StyledLink';
 
 function Header() {
   return (
     <FooterContainer>
-      <ExternalLink
+      <StyledLink
+        as="a"
         href="https://profy.dev/employers"
         fontSize="0.875rem"
         target="_blank"
       >
         profy.dev
-      </ExternalLink>
-      <InternalLink to="/">
+      </StyledLink>
+      <StyledLink to="/">
         <img src={logo} alt="logo" width="36px" height="auto" style={{ display: 'block' }} />
-      </InternalLink>
-      <InternalLink to="/terms" fontSize="0.875rem">Terms & Privacy</InternalLink>
+      </StyledLink>
+      <StyledLink to="/terms" fontSize="0.875rem">Terms & Privacy</StyledLink>
     </FooterContainer>
   );
 }
